--- a/packages/ipfs-unixfs-exporter/package.json
+++ b/packages/ipfs-unixfs-exporter/package.json
@@ -159,6 +159,10 @@
     "interface-blockstore": "^3.0.0",
     "ipfs-unixfs": "^7.0.0",
     "it-last": "^1.0.5",
+    "it-parallel": "^2.0.1",
+    "it-pipe": "^2.0.4",
+    "it-pushable": "^3.1.0",
+    "it-map": "^1.0.6",
     "multiformats": "^9.4.2",
     "uint8arrays": "^3.0.0"
   },
@@ -168,6 +172,7 @@
     "aegir": "^37.5.0",
     "blockstore-core": "^2.0.1",
     "crypto-browserify": "^3.12.0",
+    "delay": "^5.0.0",
     "ipfs-unixfs-importer": "^10.0.0",
     "it-all": "^1.0.5",
     "it-buffer-stream": "^2.0.0",