--- a/packages/ipfs-unixfs-exporter/src/utils/extract-data-from-block.js
+++ b/packages/ipfs-unixfs-exporter/src/utils/extract-data-from-block.js
@@ -16,12 +16,12 @@ function extractDataFromBlock (block, blockStart, requestedStart, requestedEnd) 
 
   if (requestedEnd >= blockStart && requestedEnd < blockEnd) {
     // If the end byte is in the current block, truncate the block to the end byte
-    block = block.slice(0, requestedEnd - blockStart)
+    block = block.subarray(0, requestedEnd - blockStart)
   }
 
   if (requestedStart >= blockStart && requestedStart < blockEnd) {
     // If the start byte is in the current block, skip to the start byte
-    block = block.slice(requestedStart - blockStart)
+    block = block.subarray(requestedStart - blockStart)
   }
 
   return block