A polishing of #237.  Uses `it-parallel` to load a whole list of children of a DAG node in parallel rather than one at a time.  Makes fetching large files much faster.

From the js-ipfs interop suite:

Before:

```console
  exchange files
    go -> js
      cat file
        ✔ go -> js: 67.1 MB (25659ms)
    go -> go2
      cat file
        ✔ go -> go2: 67.1 MB (19835ms)
    js -> go
      cat file
        ✔ js -> go: 67.1 MB (17407ms)
    js -> js2
      cat file
        ✔ js -> js2: 67.1 MB (20237ms)
```

After:

```console
  exchange files
    go -> js
      cat file
        ✔ go -> js: 67.1 MB (15328ms)
    go -> go2
      cat file
        ✔ go -> go2: 67.1 MB (18536ms)
    js -> go
      cat file
        ✔ js -> go: 67.1 MB (17211ms)
    js -> js2
      cat file
        ✔ js -> js2: 67.1 MB (12124ms)
```